Add download_result, finetune, get_models, and inference functions
