+++ requirements.txt
@@ -1,49 +1,53 @@
 # --index-url https://pypi.org/simple
 # --extra-index-url https://download.pytorch.org/whl/cu118
-# torch==2.1.1+cu118
-# torchvision==0.16.1+cu118
-# torchaudio==2.1.1+cu118
-# torchtext
+# --trusted-host pypi.org
+# torch==2.1.0+cu118
+# torchvision==0.16.0+cu118
+# torchaudio==2.1.0+cu118
 
-# --index-url https://pypi.org/simple
-# --extra-index-url https://download.pytorch.org/whl/cu116
-# torch==1.12.1+cu116
-# torchvision==0.13.1+cu116
-# torchaudio==0.12.1
-# torchtext
+--index-url https://pypi.org/simple
+--extra-index-url https://download.pytorch.org/whl/cu116
+--trusted-host pypi.org
+torch==1.12.1+cu116
+torchvision==0.13.1+cu116
+torchaudio==0.12.1+cu116
+Pillow==9.5.0  # If you are running on torch1.12, you need to install Pillow==9.5.0
 
+# Basic Packages
+lightning==2.1.2
+mlflow
+pydantic
+SQLAlchemy
+aim<4.0.0
+dvc
+dvc-gdrive
+dvc[s3]
+rich
+hydra-core
+hydra-colorlog
+hydra-optuna-sweeper
+opencv-contrib-python
+matplotlib
+rasterio
+rootutils
+pyrootutils
+autorootcwd
+shapely
+rootutils
+tensorboard
+tensorboardX
+wandb
+fire
+
+# dev packages
 pre-commit
-pytest
+ipykernel
+ipywidgets
 sh
+pytest
 
-# lightning==2.1.2
-# mlflow
-# pydantic
-# SQLAlchemy
-# aim<4.0.0
-# dvc
-# dvc[s3]
-# rich
-# hydra-core
-# hydra-colorlog
-# hydra-optuna-sweeper
-# opencv-contrib-python
-# matplotlib
-# rasterio
-# Pillow
-# pyrootutils
-# shapely
-# rootutils
-# tensorboard
-# tensorboardX
-# wandb
-# autoroot
-# autorootcwd
-# ipykernel
-# ipywidgets
-# orjsonl
-# pandas
-# pyarrow
-# fastparquet
-# sentencepiece
-# transformers
+# Captcha Resolver Dependencies
+wget
+einops
+timm==0.4.12
+gradio
