+++ src/download_result.py
@@ -1,0 +1,3 @@
+def provide_download_file():
+    file_path = "./tests.zip"
+    return file_path
